log4j with gelfappender and handle invalid pk id in queue

+++ spring-activemq-sample/src/main/java/com/blog/spring/jms/MsgCtl.java
@@ -1,9 +1,12 @@
 package com.blog.spring.jms;
 
 
+import antlr.collections.Enumerator;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.jms.core.BrowserCallback;
 import org.springframework.jms.core.JmsTemplate;
 import org.springframework.stereotype.Controller;
 import org.springframework.transaction.annotation.Propagation;
@@ -14,8 +17,11 @@ import org.springframework.web.bind.annotation.RequestParam;
 import javax.annotation.Resource;
 import javax.jms.JMSException;
 import javax.jms.Queue;
+import javax.jms.QueueBrowser;
+import javax.jms.Session;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
+import java.util.Enumeration;
 
 /**
  * Daneel Yaitskov
@@ -40,21 +46,35 @@ public class MsgCtl {
     public void send(HttpServletResponse response, @RequestParam("text") String text)
             throws JMSException, IOException {
 
-        jdbcTemplate.update("insert into msg (txt) values(?)", new Object[] { text });
-        int id = jdbcTemplate.queryForInt( "select last_insert_id()" );
+        jdbcTemplate.update("insert into msg (txt) values(?)", new Object[]{text});
+        int id = jdbcTemplate.queryForInt("select last_insert_id()");
         jmsTemplate.convertAndSend(testQueue, id);
+        logger.warn("send msg " + text);
         response.getWriter().append("ok. message '" + text + "' was sent and id = " + id);
 
     }
 
     @Transactional(propagation = Propagation.REQUIRES_NEW)
     @RequestMapping("/receive")
-    public void receive(HttpServletResponse response) throws IOException {
-        Object msg = jmsTemplate.receiveAndConvert(testQueue);
-        logger.debug("got id " + msg);
-        int id = (Integer)msg;
-        String tmsg = (String) jdbcTemplate.queryForObject("select txt from msg where id = ?", String.class, id);
-        jdbcTemplate.update("delete from msg where id = ?", id);
-        response.getWriter().append("ok. message '" + tmsg + "' is received with id = " + id);
+    public void receive(HttpServletResponse response) throws IOException, JMSException {
+        while (true) {
+            Object msg = jmsTemplate.receiveAndConvert(testQueue);
+            if (msg == null) {
+                response.getWriter().append("queue is empty");
+                logger.debug("queue {} is empty", testQueue.getQueueName());
+                return;
+            }
+            logger.debug("got id " + msg);
+            int id = (Integer) msg;
+            try {
+                String tmsg = jdbcTemplate.queryForObject("select txt from msg where id = ?", String.class, id);
+                logger.warn("send msg " + tmsg);
+                jdbcTemplate.update("delete from msg where id = ?", id);
+                response.getWriter().append("ok. message '" + tmsg + "' is received with id = " + id);
+                break;
+            } catch (EmptyResultDataAccessException e) {
+                logger.debug("id {} is missing in DB", id);
+            }
+        }
     }
 }
